# web/lib/openapi.json
{
  "components": {
    "schemas": {
...
      "ConfirmIn": {
        "properties": {
          "ai_assisted": {
            "default": false,
            "title": "Ai Assisted",
            "type": "boolean"
          },
          "confirmed_by_observer": {
            "default": false,
            "title": "Confirmed By Observer",
            "type": "boolean"
          },
          "idempotency_key": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Idempotency Key"
          },
          "method": {
            "$ref": "#/components/schemas/ObservationMethod"
          },
          "mission_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mission Id"
          },
          "node_id": {
            "title": "Node Id",
            "type": "string"
          },
          "oah_codes": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Oah Codes",
            "type": "array"
          },
          "observed_at": {
            "format": "date-time",
            "title": "Observed At",
            "type": "string"
          },
          "observer_id": {
            "title": "Observer Id",
            "type": "string"
          },
          "observer_type": {
            "title": "Observer Type",
            "type": "string"
          },
          "photo_uri": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Photo Uri"
          },
          "result": {
            "$ref": "#/components/schemas/ObservationResult"
          },
          "snap_distance_m": {
            "title": "Snap Distance M",
            "type": "number"
          },
          "unit": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Unit"
          },
          "value": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Value"
          }
        },
        "required": [
          "node_id",
          "observed_at",
          "method",
          "result",
          "observer_id",
          "observer_type",
          "snap_distance_m"
        ],
        "title": "ConfirmIn",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "LabIn": {
        "description": "A lab result is not a field observation.\n\nIt carries no `result` (it is quantitative by definition) and no snap distance:\nthe sample is tied to a node id by the chain of custody, not by a GPS fix.",
        "properties": {
          "method": {
            "$ref": "#/components/schemas/ObservationMethod"
          },
          "mission_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mission Id"
          },
          "node_id": {
            "title": "Node Id",
            "type": "string"
          },
          "oah_codes": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Oah Codes",
            "type": "array"
          },
          "observed_at": {
            "format": "date-time",
            "title": "Observed At",
            "type": "string"
          },
          "observer_id": {
            "title": "Observer Id",
            "type": "string"
          },
          "observer_type": {
            "default": "lab",
            "title": "Observer Type",
            "type": "string"
          },
          "snap_distance_m": {
            "default": 0.0,
            "title": "Snap Distance M",
            "type": "number"
          },
          "unit": {
            "title": "Unit",
            "type": "string"
          },
          "value": {
            "title": "Value",
            "type": "number"
          }
        },
        "required": [
          "node_id",
          "observed_at",
          "method",
          "value",
          "unit",
          "observer_id"
        ],
        "title": "LabIn",
        "type": "object"
      },
      "ObservationMethod": {
        "enum": [
          "citizen_visual_olfactory",
          "citizen_freetext",
          "citizen_photo",
          "test_strip",
          "sensor_turbidity",
          "sensor_conductivity",
          "sensor_normal_window",
          "field_test",
          "lab_ecoli",
          "lab_enterococci",
          "overflow_telemetry",
          "bioassessment"
        ],
        "title": "ObservationMethod",
        "type": "string"
      },
      "ObservationResult": {
        "enum": [
          "positive",
          "negative",
          "quantitative"
        ],
        "title": "ObservationResult",
        "type": "string"
      },
      "OverflowIn": {
        "properties": {
          "active": {
            "title": "Active",
            "type": "boolean"
          },
          "node_id": {
            "title": "Node Id",
            "type": "string"
          },
          "outfall_id": {
            "title": "Outfall Id",
            "type": "string"
          },
          "ts": {
            "format": "date-time",
            "title": "Ts",
            "type": "string"
          }
        },
        "required": [
          "outfall_id",
          "node_id",
          "ts",
          "active"
        ],
        "title": "OverflowIn",
        "type": "object"
      },
      "ProposeIn": {
        "properties": {
          "free_text": {
            "default": "",
            "title": "Free Text",
            "type": "string"
          },
          "lat": {
            "title": "Lat",
            "type": "number"
          },
          "lon": {
            "title": "Lon",
            "type": "number"
          },
          "observed_at": {
            "format": "date-time",
            "title": "Observed At",
            "type": "string"
          },
          "photo_uri": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Photo Uri"
          }
        },
        "required": [
          "lon",
          "lat",
          "observed_at"
        ],
        "title": "ProposeIn",
        "type": "object"
      },
      "RainfallIn": {
        "properties": {
          "antecedent_dry_h": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Antecedent Dry H"
          },
          "mm_per_h": {
            "title": "Mm Per H",
            "type": "number"
          },
          "stream": {
            "default": "live",
            "title": "Stream",
            "type": "string"
          },
          "ts": {
            "format": "date-time",
            "title": "Ts",
            "type": "string"
          }
        },
        "required": [
          "ts",
          "mm_per_h"
        ],
        "title": "RainfallIn",
        "type": "object"
      },
      "RetractIn": {
        "description": "The documented request shape for FR-10 is {event_id, reason}.",
        "properties": {
          "event_id": {
            "title": "Event Id",
            "type": "string"
          },
          "reason": {
            "title": "Reason",
            "type": "string"
          },
          "retracted_by": {
            "title": "Retracted By",
            "type": "string"
          }
        },
        "required": [
          "event_id",
          "reason",
          "retracted_by"
        ],
        "title": "RetractIn",
        "type": "object"
      },
      "SensorIn": {
        "properties": {
          "node_id": {
            "title": "Node Id",
            "type": "string"
          },
          "parameter": {
            "title": "Parameter",
            "type": "string"
          },
          "sensor_id": {
            "title": "Sensor Id",
            "type": "string"
          },
          "stream": {
            "default": "live",
            "title": "Stream",
            "type": "string"
          },
          "ts": {
            "format": "date-time",
            "title": "Ts",
            "type": "string"
          },
          "unit": {
            "title": "Unit",
            "type": "string"
          },
          "value": {
            "title": "Value",
            "type": "number"
          }
        },
        "required": [
          "sensor_id",
          "node_id",
          "ts",
          "parameter",
          "value",
          "unit"
        ],
        "title": "SensorIn",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Upstream Core API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/ingest/lab": {
      "post": {
        "description": "FR-9: lab results, including ones that arrive days later (GC-4).",
        "operationId": "lab_result_ingest_lab_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LabIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Lab Result",
        "tags": [
          "ingest"
        ]
      }
    },
    "/ingest/overflow": {
      "post": {
        "description": "FR-8: an overflow activation is both a signal and a piece of evidence.",
        "operationId": "overflow_ingest_overflow_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OverflowIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Overflow",
        "tags": [
          "ingest"
        ]
      }
    },
    "/ingest/rainfall": {
      "post": {
        "description": "FR-7: rainfall every 15 minutes, with the derived flow condition (PRD 7.9).",
        "operationId": "rainfall_ingest_rainfall_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RainfallIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Rainfall",
        "tags": [
          "ingest"
        ]
      }
    },
    "/ingest/report/confirm": {
      "post": {
        "description": "FR-1, FR-2, FR-3: record confirmed evidence, positive or negative.",
        "operationId": "confirm_ingest_report_confirm_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfirmIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Confirm",
        "tags": [
          "ingest"
        ]
      }
    },
    "/ingest/report/propose": {
      "post": {
        "description": "FR-3: AI proposes fields. Appends nothing to the log (GC-8).",
        "operationId": "propose_ingest_report_propose_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProposeIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Propose",
        "tags": [
          "ingest"
        ]
      }
    },
    "/ingest/retract": {
      "post": {
        "description": "FR-10: retraction is a new event, never a delete (GC-5).",
        "operationId": "retract_ingest_retract_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RetractIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Retract",
        "tags": [
          "ingest"
        ]
      }
    },
    "/ingest/sensor": {
      "post": {
        "description": "FR-6: raw readings land in the hypertable; derived evidence comes from sensor_job.",
        "operationId": "sensor_ingest_sensor_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SensorIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Sensor",
        "tags": [
          "ingest"
        ]
      }
    },
...
  }
}